LayoutTests/fast/scrolling/ios/ui-side-compositing-check.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../../resources/js-test-pre.js"></script>
    <script>
        if (window.testRunner)
          testRunner.dumpAsText();
        shouldBe("window.internals.isUsingUISideCompositing()", "true");
    </script>
    <script src="../../../resources/js-test-post.js"></script>
</head>
</html>
