LayoutTests/fast/scrolling/ios/ui-side-compositing-check-expected.txt

PASS window.internals.isUsingUISideCompositing() is true
PASS successfullyParsed is true

TEST COMPLETE
